parte final módulo 2

+++ src/storage/group/groupCreate.ts
@@ -1,15 +1,22 @@
-import AsyncStorage from "@react-native-async-storage/async-storage";
+import AsyncStorage from '@react-native-async-storage/async-storage';
+import { AppError } from '@utils/AppError';
 
-import { GROUP_COLLECTION } from "@storage/storageConfig";
-import { groupsGetAll } from "./groupsGetAll";
+import { groupsGetAll } from './groupsGetAll';
+import { GROUP_COLLECTION } from '@storage/storageConfig';
 
-export async function groupCreate(newGroupName: string) {
+export async function groupCreate(newGroup: string) {
   try {
-    const storagedGroups= await groupsGetAll();
+    const storedGroups = await groupsGetAll();
 
-    const storage = JSON.stringify([...storagedGroups, newGroupName])
+    const groupAlreadyExists = storedGroups.includes(newGroup);
+
+    if(groupAlreadyExists) {
+      throw new AppError('Já existe um grupo cadastrado com esse nome.')
+    }
+
+    const storage = JSON.stringify([...storedGroups, newGroup])
     await AsyncStorage.setItem(GROUP_COLLECTION, storage);
-
+    
   } catch (error) {
     throw error;
   }
